WIP Refactored updateSiteActor
